App_Engine/V2/requirements.txt
vertexai
fastapi
uvicorn
python-multipart
starlette
apscheduler
google-auth 
google-auth-oauthlib 
google-auth-httplib2
google-cloud-aiplatform
pandasai
google-generativeai
